task: Add the finished package.
This commit adds the entire "task" package of the project,
following the Clean Architecture paradigm. An extra helper.go
file is also added to store functions that are used as additional
helpers, such as for executing commands and storing byte streams
to files.

The storage used here is an in-memory array of pointers to the
Task struct.